# src/main/resources/api-docs-1.0.0.yaml
---
openapi: 3.0.0
info:
  title: КупиПодариДай
  description: API сервиса вишлистов
  version: 1.0.0
paths:
  /users/me:
    get:
      tags:
      - users
      operationId: UsersController_findOwn
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponseDto'
      security:
      - bearer: []
    patch:
      tags:
      - users
      operationId: UsersController_update
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
        required: true
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponseDto'
        "400":
          description: Ошибка валидации переданных значений
      security:
      - bearer: []
  /users/me/wishes:
    get:
      tags:
      - users
      operationId: UsersController_getOwnWishes
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wish'
      security:
      - bearer: []
  /users/{username}:
    get:
      tags:
      - users
      operationId: UsersController_findOne
      parameters:
      - name: username
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublicProfileResponseDto'
      security:
      - bearer: []
  /users/{username}/wishes:
    get:
      tags:
      - users
      operationId: UsersController_getWishes
      parameters:
      - name: username
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserWishesDto'
      security:
      - bearer: []
  /users/find:
    post:
      tags:
      - users
      operationId: UsersController_findMany
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FindUsersDto'
        required: true
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfileResponseDto'
      security:
      - bearer: []
  /wishes:
    post:
      tags:
      - wishes
      operationId: WishesController_create
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWishDto'
        required: true
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      security:
      - bearer: []
  /wishes/last:
    get:
      tags:
      - wishes
      operationId: WishesController_findLast
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wish'
      security:
      - bearer: []
  /wishes/top:
    get:
      tags:
      - wishes
      operationId: WishesController_findTop
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wish'
      security:
      - bearer: []
  /wishes/{id}:
    get:
      tags:
      - wishes
      operationId: WishesController_findOne
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wish'
      security:
      - bearer: []
    delete:
      tags:
      - wishes
      operationId: WishesController_removeOne
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wish'
      security:
      - bearer: []
    patch:
      tags:
      - wishes
      operationId: WishesController_update
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWishDto'
        required: true
      responses:
        "200":
          description: ""
      security:
      - bearer: []
  /wishes/{id}/copy:
    post:
      tags:
      - wishes
      operationId: WishesController_copyWish
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      security:
      - bearer: []
  /offers:
    get:
      tags:
      - offers
      operationId: OffersController_findAll
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Offer'
      security:
      - bearer: []
    post:
      tags:
      - offers
      operationId: OffersController_create
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOfferDto'
        required: true
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      security:
      - bearer: []
  /offers/{id}:
    get:
      tags:
      - offers
      operationId: OffersController_findOne
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
      security:
      - bearer: []
  /wishlistlists:
    get:
      tags:
      - wishlistlists
      operationId: WishlistsController_findAll
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wishlist'
    post:
      tags:
      - wishlistlists
      operationId: WishlistsController_create
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWishlistDto'
        required: true
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wishlist'
      security:
      - bearer: []
  /wishlistlists/{id}:
    get:
      tags:
      - wishlistlists
      operationId: WishlistsController_findOne
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wishlist'
      security:
      - bearer: []
    delete:
      tags:
      - wishlistlists
      operationId: WishlistsController_removeOne
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wishlist'
      security:
      - bearer: []
    patch:
      tags:
      - wishlistlists
      operationId: WishlistsController_update
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWishlistDto'
        required: true
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wishlist'
      security:
      - bearer: []
  /signin:
    post:
      tags:
      - auth
      operationId: AuthController_login
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigninUserDto'
        required: true
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SigninUserResponseDto'
        "401":
          description: Некорректная пара логин и пароль
  /signup:
    post:
      tags:
      - auth
      operationId: AuthController_signup
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
        required: true
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupUserResponseDto'
        "409":
          description: Пользователь с таким email или username уже зарегистрирован
components:
  schemas:
    UserProfileResponseDto:
      required:
      - about
      - avatar
      - createdAt
      - email
      - id
      - updatedAt
      - username
      type: object
      properties:
        id:
          type: number
          example: 5
        username:
          maxLength: 64
          minLength: 1
          type: string
          example: user
        about:
          maxLength: 200
          minLength: 1
          type: string
          example: Пока ничего не рассказал о себе
        avatar:
          type: string
          example: https://i.pravatar.cc/300
        email:
          type: string
          example: user@yandex.ru
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateUserDto:
      type: object
      properties:
        username:
          maxLength: 64
          minLength: 1
          type: string
          example: exampleuser
        about:
          maxLength: 200
          minLength: 0
          type: string
          example: exampleuser
        avatar:
          type: string
          example: https://i.pravatar.cc/150?img=3
        email:
          type: string
          example: user@yandex.ru
        password:
          minLength: 2
          type: string
          example: somestrongpassword
    Offer:
      required:
      - amount
      - createdAt
      - hidden
      - id
      - item
      - updatedAt
      - user
      type: object
      properties:
        id:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        item:
          $ref: '#/components/schemas/Wish'
        amount:
          type: number
        hidden:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
    WishPartial:
      type: object
      properties:
        id:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        name:
          maxLength: 250
          minLength: 1
          type: string
        link:
          type: string
        image:
          type: string
        price:
          minimum: 1
          type: number
        raised:
          minimum: 1
          type: number
        copied:
          type: number
        description:
          maxLength: 1024
          minLength: 1
          type: string
    Wish:
      required:
      - copied
      - createdAt
      - description
      - id
      - image
      - link
      - name
      - offers
      - owner
      - price
      - raised
      - updatedAt
      type: object
      properties:
        id:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        name:
          maxLength: 250
          minLength: 1
          type: string
        link:
          type: string
        image:
          type: string
        price:
          minimum: 1
          type: number
        raised:
          minimum: 1
          type: number
        copied:
          type: number
        description:
          maxLength: 1024
          minLength: 1
          type: string
        owner:
          $ref: '#/components/schemas/UserPublicProfileResponseDto'
        offers:
          type: array
          items:
            $ref: '#/components/schemas/Offer'
    Wishlist:
      required:
      - createdAt
      - id
      - image
      - items
      - name
      - owner
      - updatedAt
      type: object
      properties:
        id:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        name:
          maxLength: 250
          minLength: 0
          type: string
        image:
          type: string
        owner:
          $ref: '#/components/schemas/UserPublicProfileResponseDto'
        items:
          type: array
          items:
            $ref: '#/components/schemas/WishPartial'
    User:
      required:
      - about
      - avatar
      - createdAt
      - email
      - id
      - offers
      - updatedAt
      - username
      - wishes
      - wishlists
      type: object
      properties:
        id:
          type: number
          example: 5
        username:
          maxLength: 64
          minLength: 1
          type: string
          example: user
        about:
          maxLength: 200
          minLength: 1
          type: string
          example: Пока ничего не рассказал о себе
        avatar:
          type: string
          example: https://i.pravatar.cc/300
        email:
          type: string
          example: user@yandex.ru
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        wishes:
          type: array
          items:
            $ref: '#/components/schemas/Wish'
        offers:
          type: array
          items:
            $ref: '#/components/schemas/Wish'
        wishlists:
          type: array
          items:
            $ref: '#/components/schemas/Wishlist'
    UserPublicProfileResponseDto:
      required:
      - about
      - avatar
      - createdAt
      - id
      - updatedAt
      - username
      type: object
      properties:
        id:
          type: number
          example: 5
        username:
          maxLength: 64
          minLength: 1
          type: string
          example: user
        about:
          maxLength: 200
          minLength: 1
          type: string
          example: Пока ничего не рассказал о себе
        avatar:
          type: string
          example: https://i.pravatar.cc/300
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserWishesDto:
      required:
      - copied
      - createdAt
      - description
      - id
      - image
      - link
      - name
      - offers
      - price
      - raised
      - updatedAt
      type: object
      properties:
        id:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        name:
          maxLength: 250
          minLength: 1
          type: string
        link:
          type: string
        image:
          type: string
        price:
          minimum: 1
          type: number
        raised:
          minimum: 1
          type: number
        copied:
          type: number
        description:
          maxLength: 1024
          minLength: 1
          type: string
        offers:
          type: array
          items:
            $ref: '#/components/schemas/Offer'
    FindUsersDto:
      required:
      - query
      type: object
      properties:
        query:
          type: string
          description: username или email
          example: some@ya.ru
    CreateWishDto:
      required:
      - description
      - image
      - link
      - name
      - price
      type: object
      properties:
        name:
          maxLength: 250
          minLength: 1
          type: string
        link:
          type: string
        image:
          type: string
        price:
          minimum: 1
          type: number
        description:
          type: string
    UpdateWishDto:
      type: object
      properties: {}
    CreateOfferDto:
      required:
      - amount
      - itemId
      type: object
      properties:
        amount:
          minimum: 1
          type: number
        hidden:
          type: boolean
        itemId:
          type: number
    CreateWishlistDto:
      type: object
      properties:
        name:
          type: string
          example: Мой вишлист
        image:
          type: string
          example: https://i.pravatar.cc/150?img=3
        itemsId:
          type: array
          example:
          - 1
          items:
            type: number
    UpdateWishlistDto:
      type: object
      properties:
        name:
          type: string
          example: Мой вишлист
        image:
          type: string
          example: https://i.pravatar.cc/150?img=3
        itemsId:
          type: array
          example:
          - 1
          items:
            type: number
    SigninUserDto:
      required:
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 64
          minLength: 1
          type: string
          example: exampleuser
        password:
          minLength: 2
          type: string
          example: somestrongpassword
    SigninUserResponseDto:
      required:
      - access_token
      type: object
      properties:
        access_token:
          type: string
          description: JWT-токен
    CreateUserDto:
      required:
      - email
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 64
          minLength: 1
          type: string
          example: exampleuser
        about:
          maxLength: 200
          minLength: 0
          type: string
          example: exampleuser
        avatar:
          type: string
          example: https://i.pravatar.cc/150?img=3
        email:
          type: string
          example: user@yandex.ru
        password:
          minLength: 2
          type: string
          example: somestrongpassword
    SignupUserResponseDto:
      required:
      - about
      - avatar
      - createdAt
      - email
      - id
      - updatedAt
      - username
      type: object
      properties:
        id:
          type: number
          example: 5
        username:
          maxLength: 64
          minLength: 1
          type: string
          example: user
        about:
          maxLength: 200
          minLength: 1
          type: string
          example: Пока ничего не рассказал о себе
        avatar:
          type: string
          example: https://i.pravatar.cc/300
        email:
          type: string
          example: user@yandex.ru
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT